% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentError_DoLatentErrorRun.R
\name{LatentErrorRun}
\alias{LatentErrorRun}
\title{MainFunction}
\usage{
LatentErrorRun(Yobs, AnalyzeFactors, factorizePolicies = F)
}
\arguments{
\item{x}{The input}
}
\value{
\code{x} Returns this.
}
\description{
Implements fancy analysis
}
\details{
Details.
}
\examples{

# Comment here
#x <- MainFunction(x)
}
