% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentError_DoSensitivity.R
\name{LatentErrorSensitivity}
\alias{LatentErrorSensitivity}
\title{MainFunction}
\usage{
LatentErrorSensitivity(x)
}
\arguments{
\item{x}{The input}
}
\value{
\code{x} Returns this.
}
\description{
Implements fancy analysis
}
\details{
Details.
}
\examples{

# Comment here
#x <- MainFunction(x)
}
