% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentError_DoOneRun.R
\name{LatentOneRun}
\alias{LatentOneRun}
\title{MainFunction}
\usage{
LatentOneRun(
  Yobs,
  ObservablesMat,
  ObservablesGroupings = colnames(ObservablesMat),
  MakeObservablesGroupings = F
)
}
\arguments{
\item{x}{The input}
}
\value{
\code{x} Returns this.
}
\description{
Implements fancy analysis
}
\details{
Details.
}
\examples{

# Comment here
#x <- MainFunction(x)
}
