% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentError_DoBootPartition.R
\name{LatentRun}
\alias{LatentRun}
\title{LatentErrorBoot}
\usage{
LatentRun(
  Yobs,
  ObservablesMat,
  ObservablesGroupings = colnames(ObservablesMat),
  MakeObservablesGroupings = F,
  nBoot = 32L,
  nPartition = 10L,
  ReturnIntermediaries = T
)
}
\arguments{
\item{x}{The input}
}
\value{
\code{x} Returns this.
}
\description{
Implements fancy analysis
}
\details{
Details.
}
\examples{

# Comment here
#x <- LatentErrorBoot(x)
}
